Macro-ize interpolate/set_strategy/DynInterpolator across Interp1D/2D/3D/ND
- interpolate_impl!(): the fixed-N loop-based `interpolate` body was already
  identical between Interp2D and Interp3D; Interp1D had its own hand-unrolled
  N=1 version doing the same thing without the loop or Vec. Unified onto the
  same macro (1D/2D/3D only; InterpND has no fixed N to loop over and keeps
  its hand-written interpolate).
- set_strategy_box_impl!/set_strategy_enum_impl!: the two set_strategy
  overloads (Box<dyn Strategy*D<D>> and Strategy*DEnum) were identical modulo
  type names across all four dimensionalities, including InterpND.
- dyn_interpolator_impl!: same story for the DynInterpolator impl on each
  owned type.

Doc comments with self-referential doc links are preserved via
concat!/stringify!, same approach as the previous traits.rs macro.
